# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API Sistema de Gestión de Solicitudes Académicas
  description: |
    API RESTful para la gestión de solicitudes académicas del
    Programa de Ingeniería de Sistemas y Computación - Universidad del Quindío.

    **Dominio principal:** Gestión de Solicitudes Académicas
    **Agregado principal:** Solicitud

    ## Ciclo de vida de una Solicitud
    ```
    REGISTRADA → CLASIFICADA → EN_ATENCION → ATENDIDA → CERRADA
    ```
    Cada transición corresponde a una operación del dominio expuesta como endpoint.

    ## Autenticación
    Todos los endpoints requieren un token JWT en el header `Authorization`.
    Formato: `Bearer <token>`

    Los roles disponibles y sus permisos son:
    - **ESTUDIANTE** → puede crear solicitudes y consultar las propias
    - **DOCENTE / ADMINISTRATIVO** → puede clasificar, asignar y atender
    - **DIRECTIVO** → acceso completo, incluyendo cerrar solicitudes

  version: 1.0.0
  contact:
    name: Programa de Ingeniería de Sistemas
    email: sistemas@uniquindio.edu.co
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
security:
  - bearerAuth: []
paths:
  /api/solicitudes:
    post:
      tags:
        - Solicitudes
      summary: Registrar una nueva solicitud
      description: |
        Crea una nueva solicitud en estado **REGISTRADA**.
        Mapea a la operación del dominio: `CrearSolicitudUseCase`
        El código se genera automáticamente — no debe enviarse en el body.

        **Roles permitidos:** ESTUDIANTE, DOCENTE, ADMINISTRATIVO, DIRECTIVO
      operationId: crearSolicitud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSolicitudRequest'
            example:
              tipo: REGISTRAR_ASIGNATURA
              descripcion: "Necesito registrar la asignatura Programación Avanzada para este semestre"
              documentoSolicitante: "1094123456"
      responses:
        '201':
          description: Solicitud registrada exitosamente
          headers:
            Location:
              description: URL de la solicitud creada
              schema:
                type: string
                example: /api/solicitudes/001
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 400
                error: "Bad Request"
                mensaje: "La descripción debe tener entre 10 y 500 caracteres"
                path: "/api/solicitudes"
        '401':
          description: No autenticado — token ausente o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes"
        '403':
          description: Sin permisos para realizar esta operación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "No tiene permisos para crear solicitudes"
                path: "/api/solicitudes"
        '404':
          description: Solicitante no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró un usuario con el documento: 1094123456"
                path: "/api/solicitudes"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes"

    get:
      tags:
        - Solicitudes
      summary: Listar solicitudes
      description: |
        Obtiene la lista de solicitudes. Se puede filtrar por estado.
        Mapea a: `ConsultarSolicitudesPorEstadoUseCase`

        **Roles permitidos:** DOCENTE, ADMINISTRATIVO, DIRECTIVO
      operationId: listarSolicitudes
      parameters:
        - name: estado
          in: query
          description: Filtrar por estado de la solicitud
          required: false
          schema:
            $ref: '#/components/schemas/EstadoDeSolicitud'
      responses:
        '200':
          description: Lista de solicitudes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SolicitudResumenResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes"
        '403':
          description: Sin permisos para listar solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "El rol ESTUDIANTE no puede listar todas las solicitudes"
                path: "/api/solicitudes"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes"
  /api/solicitudes/{codigo}:
    get:
      tags:
        - Solicitudes
      summary: Obtener detalle de una solicitud
      description: |
        Retorna la información completa de una solicitud por su código.

        **Roles permitidos:** ESTUDIANTE (solo las propias), DOCENTE, ADMINISTRATIVO, DIRECTIVO
      operationId: obtenerSolicitud
      parameters:
        - name: codigo
          in: path
          required: true
          description: Código único de la solicitud (ej. 001, 042)
          schema:
            type: string
            example: "001"
      responses:
        '200':
          description: Solicitud encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes/001"
        '403':
          description: Sin permisos para ver esta solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "No tiene permisos para consultar esta solicitud"
                path: "/api/solicitudes/001"
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró una solicitud con el código: 001"
                path: "/api/solicitudes/001"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes/001"
  /api/solicitudes/{codigo}/clasificar:
    put:
      tags:
        - Solicitudes
      summary: Clasificar una solicitud
      description: |
        Asigna una prioridad a la solicitud y la pasa a estado **CLASIFICADA**.
        Mapea a: `CambiarEstadoUseCase.clasificar()`

        **Precondiciones:**
        - La solicitud debe estar en estado `REGISTRADA`

        **Roles permitidos:** DOCENTE, ADMINISTRATIVO, DIRECTIVO
      operationId: clasificarSolicitud
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClasificarSolicitudRequest'
            example:
              prioridad: ALTO
              justificacion: "Inicio de semestre, estudiante requiere acceso urgente a la asignatura"
      responses:
        '200':
          description: Solicitud clasificada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Estado inválido para clasificar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 400
                error: "Bad Request"
                mensaje: "Solo se puede clasificar una solicitud registrada"
                path: "/api/solicitudes/001/clasificar"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes/001/clasificar"
        '403':
          description: Sin permisos para clasificar solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "El rol ESTUDIANTE no puede clasificar solicitudes"
                path: "/api/solicitudes/001/clasificar"
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró una solicitud con el código: 001"
                path: "/api/solicitudes/001/clasificar"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes/001/clasificar"
  /api/solicitudes/{codigo}/asignar:
    put:
      tags:
        - Solicitudes
      summary: Asignar responsable a una solicitud
      description: |
        Designa un usuario como responsable de atender la solicitud.
        La solicitud pasa a estado **EN_ATENCION**.
        Mapea a: `AsignarResponsableUseCase`

        **Precondiciones:**
        - La solicitud debe estar en estado `CLASIFICADA`
        - El responsable no puede tener rol `ESTUDIANTE`

        **Roles permitidos:** ADMINISTRATIVO, DIRECTIVO
      operationId: asignarResponsable
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarResponsableRequest'
            example:
              documentoResponsable: "987654321"
      responses:
        '200':
          description: Responsable asignado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Estado inválido o responsable no válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 400
                error: "Bad Request"
                mensaje: "Un estudiante no puede ser responsable"
                path: "/api/solicitudes/001/asignar"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes/001/asignar"
        '403':
          description: Sin permisos para asignar responsables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "Solo ADMINISTRATIVO o DIRECTIVO pueden asignar responsables"
                path: "/api/solicitudes/001/asignar"
        '404':
          description: Solicitud o responsable no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró un usuario con el documento: 987654321"
                path: "/api/solicitudes/001/asignar"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes/001/asignar"
  /api/solicitudes/{codigo}/atender:
    patch:
      tags:
        - Solicitudes
      summary: Marcar una solicitud como atendida
      description: |
        Registra que la solicitud fue atendida. Pasa a estado **ATENDIDA**.
        Mapea a: `CambiarEstadoUseCase.atender()`

        **Precondiciones:**
        - La solicitud debe estar en estado `EN_ATENCION`

        **Roles permitidos:** DOCENTE, ADMINISTRATIVO, DIRECTIVO
      operationId: atenderSolicitud
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtenderSolicitudRequest'
            example:
              observacion: "Se realizó el registro de la asignatura exitosamente en el sistema"
      responses:
        '200':
          description: Solicitud marcada como atendida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Estado inválido para atender
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 400
                error: "Bad Request"
                mensaje: "Solo se puede atender una solicitud en atención"
                path: "/api/solicitudes/001/atender"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes/001/atender"
        '403':
          description: Sin permisos para atender solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "El rol ESTUDIANTE no puede atender solicitudes"
                path: "/api/solicitudes/001/atender"
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró una solicitud con el código: 001"
                path: "/api/solicitudes/001/atender"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes/001/atender"
  /api/solicitudes/{codigo}/cerrar:
    put:
      tags:
        - Solicitudes
      summary: Cerrar una solicitud
      description: |
        Finaliza el ciclo de vida de la solicitud. Pasa a estado **CERRADA** (estado final).
        Mapea a: `CerrarSolicitudUseCase`

        **Precondiciones:**
        - La solicitud debe estar en estado `ATENDIDA`

        **Roles permitidos:** ADMINISTRATIVO, DIRECTIVO
      operationId: cerrarSolicitud
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarSolicitudRequest'
            example:
              observacion: "Solicitud resuelta satisfactoriamente. Estudiante notificado."
      responses:
        '200':
          description: Solicitud cerrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Estado inválido para cerrar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 400
                error: "Bad Request"
                mensaje: "Solo se puede cerrar una solicitud atendida"
                path: "/api/solicitudes/001/cerrar"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes/001/cerrar"
        '403':
          description: Sin permisos para cerrar solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "Solo ADMINISTRATIVO o DIRECTIVO pueden cerrar solicitudes"
                path: "/api/solicitudes/001/cerrar"
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró una solicitud con el código: 001"
                path: "/api/solicitudes/001/cerrar"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes/001/cerrar"
  /api/solicitudes/{codigo}/historial:
    get:
      tags:
        - Solicitudes
      summary: Obtener historial de una solicitud
      description: |
        Retorna todos los registros del historial de la solicitud en orden cronológico.
        Accede directamente al historial del agregado `Solicitud`.

        **Roles permitidos:** ESTUDIANTE (solo las propias), DOCENTE, ADMINISTRATIVO, DIRECTIVO
      operationId: obtenerHistorial
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Historial obtenido exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistroHistorialResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 401
                error: "Unauthorized"
                mensaje: "Token JWT ausente o expirado"
                path: "/api/solicitudes/001/historial"
        '403':
          description: Sin permisos para ver este historial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 403
                error: "Forbidden"
                mensaje: "No tiene permisos para consultar el historial de esta solicitud"
                path: "/api/solicitudes/001/historial"
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 404
                error: "Not Found"
                mensaje: "No se encontró una solicitud con el código: 001"
                path: "/api/solicitudes/001/historial"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-03-15T10:30:00"
                status: 500
                error: "Internal Server Error"
                mensaje: "Ocurrió un error inesperado. Por favor intente más tarde."
                path: "/api/solicitudes/001/historial"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido al autenticarse en el sistema.
        Incluirlo en cada petición como: `Authorization: Bearer <token>`
  schemas:
    CrearSolicitudRequest:
      type: object
      required:
        - tipo
        - descripcion
        - documentoSolicitante
      properties:
        tipo:
          $ref: '#/components/schemas/TipoDeSolicitud'
        descripcion:
          type: string
          minLength: 10
          maxLength: 500
          description: Descripción detallada de la solicitud
        documentoSolicitante:
          type: string
          description: Número de documento del usuario que realiza la solicitud
    ClasificarSolicitudRequest:
      type: object
      required:
        - prioridad
        - justificacion
      properties:
        prioridad:
          $ref: '#/components/schemas/PrioridadDeSolicitud'
        justificacion:
          type: string
          minLength: 10
          maxLength: 500
...
    AtenderSolicitudRequest:
      type: object
      required:
        - observacion
      properties:
        observacion:
          type: string
          maxLength: 500
...
    SolicitudDetalleResponse:
      type: object
      properties:
        codigo:
          type: string
          example: "001"
        tipo:
          $ref: '#/components/schemas/TipoDeSolicitud'
        descripcion:
          type: string
        estado:
          $ref: '#/components/schemas/EstadoDeSolicitud'
        prioridad:
          $ref: '#/components/schemas/PrioridadDeSolicitud'
          nullable: true
        solicitante:
          $ref: '#/components/schemas/UsuarioResumenResponse'
        responsable:
          $ref: '#/components/schemas/UsuarioResumenResponse'
          nullable: true
        fechaCreacion:
          type: string
          format: date-time
        totalRegistrosHistorial:
          type: integer
...
    EstadoDeSolicitud:
      type: string
      enum:
        - REGISTRADA
        - CLASIFICADA
        - EN_ATENCION
        - ATENDIDA
        - CERRADA
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        mensaje:
          type: string
          example: "La descripción es obligatoria"
        path:
          type: string
          example: "/api/solicitudes"